test(dspark): two files instead of three, and skip without aiter
CI runs the unit tests with no aiter installed. The compiled draft now
imports support_torch_compile at module load, which reaches aiter through
graph_marker, so test_dspark.py died in collection rather than skipping.
Guard it with importorskip, the same way the other aiter-dependent test
modules do.

Consolidation, no coverage lost:
- test_dspark_ragged_bounds.py folded into test_dspark_scheduler.py; both
  cover the same ell/ragged sizing, and it was the only new test file.
- the four batch-padding cases collapse into one parametrized test. They
  shared ~40 lines of stub scaffolding each to assert one rule: pad only
  a dummy run at B==1. The table now states that rule in four rows.

82 tests pass; with aiter absent the module skips.
